% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{apply_thinning}
\alias{apply_thinning}
\title{Apply Thinning}
\usage{
apply_thinning(results, thinning_factor)
}
\arguments{
\item{results}{A list of MCMC chain results.}

\item{thinning_factor}{The factor by which to thin the results (positive integer).}
}
\value{
A list of results with thinning applied.
}
\description{
Apply Thinning
}
\examples{
thinned_results <- apply_thinning(mcmc_results, thinning_factor = 2)

}
