% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPenetrance.R
\name{mhChain}
\alias{mhChain}
\title{Execution of a Single Chain in Metropolis-Hastings}
\usage{
mhChain(
  seed,
  n_iter,
  chain_id,
  data,
  save_interval,
  m1,
  m2,
  max_age,
  shift_prior_min,
  shift_prior_max,
  p0,
  q1,
  q2,
  g1,
  g2,
  PanelPRODatabase
)
}
\arguments{
\item{seed}{Seed value for random number generation.}

\item{n_iter}{Number of iterations for the chain.}

\item{chain_id}{Identifier for the chain.}

\item{data}{List of families data.}

\item{m1}{parameter for the beta distribution for the median.}

\item{m2}{parameter for the beta distribution fo the median.}

\item{max_age}{Maximum age to be considered.}

\item{shift_prior_min}{Minimum possible value for the shift parameter.}

\item{shift_prior_max}{Maximum possible value for the shift parameter.}

\item{p0}{baseline lifetime risk.}

\item{q1}{parameter of the beta distribution for the first quartile.}

\item{q2}{parameter of the beta distribution for the first quartile.}

\item{g1}{parameter of the beta distribution for the shift.}

\item{g2}{parameter of the beta distribution for the shift.}
}
\value{
A list with samples and rejection rate.
}
\description{
Execution of a Single Chain in Metropolis-Hastings
}
