% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikelihood.R
\name{mhLogLikelihood}
\alias{mhLogLikelihood}
\title{Log-Likelihood Calculation for Metropolis-Hastings}
\usage{
mhLogLikelihood(paras, families, max_age, PanelPRODatabase)
}
\arguments{
\item{paras}{Vector of parameters for which likelihood is to be computed.}

\item{data}{List of families data.}
}
\value{
Log-likelihood value.
}
\description{
Log-Likelihood Calculation for Metropolis-Hastings
}
