% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPenetrance.R
\name{PenEstim}
\alias{PenEstim}
\title{Bayesian Inference using Independent Metropolis-Hastings for Penetrance Estimation}
\usage{
PenEstim(
  data,
  n_chains,
  n_iter_per_chain,
  max_age = 94,
  proposal_dist,
  proposal_params
)
}
\arguments{
\item{data}{List of families data.}

\item{n_chains}{Number of chains for parallel computation.}

\item{n_iter_per_chain}{Number of iterations for each chain.}

\item{max_age}{Maximum age to be considered.}

\item{proposal_dist}{List of the statistical distribution for the proposal distributions for the median, first quartile, shift and asymptote parameters.}

\item{proposal_params}{List of the parameters for the distributions of the proposal.}
}
\value{
A list containing results for each chain.
}
\description{
This function employs a Bayesian approach for penetrance estimation, utilizing the
Independent Metropolis-Hastings algorithm. It leverages parallel computing and requires
the \code{stats}, \code{parallel}, and \code{PPP} packages.
}
