% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPenetrance.R
\name{mhChain}
\alias{mhChain}
\title{Execution of a Single Chain in Metropolis-Hastings}
\usage{
mhChain(
  seed,
  n_iter,
  chain_id,
  data,
  proposal_params,
  max_age,
  PanelPRODatabase
)
}
\arguments{
\item{seed}{Seed value for random number generation.}

\item{n_iter}{Number of iterations for the chain.}

\item{chain_id}{Identifier for the chain.}

\item{data}{List of families data.}

\item{proposal_params}{List of the parameters for the distributions of the proposal.}

\item{max_age}{Maximum age to be considered.}

\item{proposal_dist}{List of the statistical distribution for the proposal distributions for the median, first quartile, shift and asymptote parameters.}
}
\value{
A list with samples and rejection rate.
}
\description{
Execution of a Single Chain in Metropolis-Hastings
}
