% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPenetrance.R
\name{PenEstim}
\alias{PenEstim}
\title{Bayesian Inference using Independent Metropolis-Hastings for Penetrance Estimation}
\usage{
PenEstim(
  data,
  n_chains,
  n_iter_per_chain,
  save_interval = 200,
  max_age = 94,
  shift_prior_min = 0,
  shift_prior_max = 25,
  p0 = 0.15,
  m1 = 2,
  m2 = 2,
  q1 = 6,
  q2 = 3,
  g1 = 9,
  g2 = 1
)
}
\arguments{
\item{data}{List of families data.}

\item{n_chains}{Number of chains for parallel computation.}

\item{n_iter_per_chain}{Number of iterations for each chain.}

\item{save_interval}{Interval after which states should be saved.}

\item{max_age}{Maximum age to be considered.}

\item{shift_prior_min}{Minimum possible value for the shift parameter.}

\item{shift_prior_max}{Maximum possible value for the shift parameter.}

\item{p0}{baseline lifetime risk.}

\item{m1}{parameter for the beta distribution for the median.}

\item{m2}{parameter for the beta distribution fo the median.}

\item{q1}{parameter of the beta distribution for the first quartile.}

\item{q2}{parameter of the beta distribution for the first quartile.}

\item{g1}{parameter of the beta distribution for the shift.}

\item{g2}{parameter of the beta distribution for the shift.}
}
\value{
A list containing results for each chain.
}
\description{
This function employs a Bayesian approach for penetrance estimation, utilizing the
Independent Metropolis-Hastings algorithm. It leverages parallel computing and requires
the \code{stats}, \code{parallel}, and \code{PPP} packages.
}
